requirements.txt
tensorflow>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
xgboost>=1.7.0
Pillow>=10.0.0
gradio>=4.0.0
kaggle>=1.5.16
jupyter>=1.0.0
